refact: lower layer cannot know higher layers
