+++ vaultscan/core/engines/base.py
@@ -1,7 +1,9 @@
-from typing import List, Dict, Optional
-from dataclasses import dataclass, field
+from typing import List, Optional
+from dataclasses import dataclass
 from enum import Enum
 from abc import ABC, abstractmethod
+
+from vaultscan.repositories.vault.base import BaseVaultConfig
 
 
 @dataclass
@@ -9,29 +11,6 @@ class Secret:
     vault: str
     name: str
     value: Optional[str] = ''
-
-
-class VaultStatus(Enum):
-    ENABLED = 'enabled'
-    DISABLED = 'disabled'
-
-
-@dataclass
-class BaseVaultConfig:
-    alias: str
-    type: str = field(init = False)  # Each subclass should set this
-    # FIX IT: Change it to VaultStatus
-    status: str
-
-    def __post_init__(self):
-        if not hasattr(self, "type"):
-            raise NotImplementedError("Subclasses must define a 'type attribute")
-
-    def to_dict(self) -> Dict:
-        return self.__dict__
-
-    @classmethod
-    def from_dict(cls, content: Dict) -> 'BaseVaultConfig': pass
 
 
 class FilterType(Enum):
