+++ vaultscan/repositories/config/in_memory.py
@@ -1,6 +1,5 @@
 from typing import List, Dict
 
-
 from vaultscan.repositories.config.base import ConfigRepository, Config
